backend/requirements.txt
torch
open_clip_torch
"fastapi[standard]